move IAgentProps to interfaces folder
