+++ src/interfaces/IAgentProps.tsx
@@ -1,0 +1,7 @@
+
+export interface IAgentProps {
+  dealerHasControl: Boolean;
+  gameStatus: Number;
+  handValue: any;
+  playerKey: string;
+}
